Move HTTPService and ConfigService into utils package
