+++ utils/config.go
@@ -852,3 +852,10 @@ func ValidateLocales(cfg *model.Config) *model.AppError {
 
 	return err
 }
+
+// An interface representing something that contains a Config, such as the app.App struct
+type ConfigService interface {
+	Config() *model.Config
+	AddConfigListener(func(old, current *model.Config)) string
+	RemoveConfigListener(string)
+}
